data_creator/requirements.txt
fastapi==0.116.0
pyarrow==20.0.0
yfinance==0.2.65
uvicorn==0.35.0
pandas==2.3.1
kafka-python==2.2.15
pydantic==2.11.7
boto3==1.39.4
